deployments/whisper_transcribe/requirements.txt
openai-whisper==20231106
requests==2.31.0